+++ R/CIDists.R
@@ -31,47 +31,144 @@
 #' @keywords htest
 #'
 #' @examples
+#' ## All types at once
+#' binCI(7,20)
+#' 
+#' ## Individual types
 #' binCI(7,20,type="wilson")
 #' binCI(7,20,type="exact")
 #' binCI(7,20,type="asymptotic")
 #' binCI(7,20,type="asymptotic",verbose=TRUE)
 #' 
-#' ## Demonstrates using all types at once
-#' binCI(7,20,type="all")
-#' binCI(7,20,type="all",verbose=TRUE)
+#' ## Multiple types
+#' binCI(7,20,type=c("exact","asymptotic"))
+#' binCI(7,20,type=c("exact","asymptotic"),verbose=TRUE)
 #' 
-#' ## Demonstrates use with multiple inputs
-#' binCI(c(7,10),c(20,30))
-#' binCI(c(7,10),c(20,30),verbose=TRUE)
+#' ## Use with multiple inputs
+#' binCI(c(7,10),c(20,30),type="wilson")
+#' binCI(c(7,10),c(20,30),type="wilson",verbose=TRUE)
 #'
 #' @export
-binCI <- function(x,n,conf.level=0.95,type=c("wilson","exact","asymptotic","all"),
+binCI <- function(x,n,conf.level=0.95,type=c("wilson","exact","asymptotic"),
                   verbose=FALSE) {
-  type <- match.arg(type)
-  if (!is.vector(x)) STOP("'x' must be a single numeric or a vector of numerics.")
-  if (!is.numeric(x)) STOP("'x' must be numeric.")
-  if (!is.numeric(n)) STOP("'n' must be numeric.")
+  type <- match.arg(type,several.ok=TRUE)
+  if (!is.vector(x)) STOP("'x' must be a single or vector of whole numbers.")
+  if (!is.vector(n)) STOP("'n' must be a single or vector of whole numbers.")
+  if (!is.numeric(x)) STOP("'x' must be whole numbers.")
+  if (!is.numeric(n)) STOP("'n' must be whole numbers.")
+  if (!all(is.wholenumber(x))) STOP("'x' must be whole numbers.")
+  if (!all(is.wholenumber(n))) STOP("'n' must be whole numbers.")
   if (any(x<0)) STOP("'x' must be non-negative.")
   if (any(n<0)) STOP("'n' must be non-negative.")
   if (any(x>n)) STOP("'x' must not be greater than 'n'.")
-  switch(type,
-         all = {
-           if (length(x)>1) STOP("'type=all' does not work with vector inputs.")
-           res <- rbind(epitools::binom.exact(x,n,conf.level),
-                        epitools::binom.wilson(x,n,conf.level),
-                        epitools::binom.approx(x,n,conf.level))
-           rownames(res) <- c("Exact","Wilson","Asymptotic")
-           },
-         exact = { res <- epitools::binom.exact(x,n,conf.level) },
-         wilson = { res <- epitools::binom.wilson(x,n,conf.level) },
-         asymptotic = { res <- epitools::binom.approx(x,n,conf.level) })
-  # relabel CI columns, convert to matrix, drop "conf.level" column (6th)
+  if (conf.level<=0 | conf.level>=1) STOP("'conf.level' must be between 0 and 1.")
+  ## Handle differently depending on number of xs given
+  if (length(x)>1) {
+    if (length(type)>1) {
+      type <- type[1]
+      WARN("Can't use multiple 'type's with multiple 'x's. Used only '",type,"'.")
+    }
+    switch(type,
+           exact = { res <- epitools::binom.exact(x,n,conf.level) },
+           wilson = { res <- epitools::binom.wilson(x,n,conf.level) },
+           asymptotic = { res <- epitools::binom.approx(x,n,conf.level) })
+  } else {
+    res <- rbind(epitools::binom.exact(x,n,conf.level),
+                 epitools::binom.wilson(x,n,conf.level),
+                 epitools::binom.approx(x,n,conf.level))
+    rownames(res) <- c("Exact","Wilson","Asymptotic")
+    # reduce to selected types
+    res <- res[rownames(res) %in% capFirst(type),]
+  }
+  # relabel CI columns, convert to matrix, drop unwanted columns
   names(res)[which(names(res) %in% c("lower","upper"))] <- iCILabel(conf.level)
   res <- as.matrix(res[,-6])
-  # remove rownnames if not type="all"
-  if (type!="all") rownames(res) <- rep("",nrow(res))
   # return everything if verbose=TRUE, otherwise just CI
-  if (!verbose) res <- res[,4:5,drop=FALSE]
+  if (!verbose) {
+    res <- res[,4:5,drop=FALSE]
+    # remove rownname if only one type selected and not verbose
+    if (length(type)==1) rownames(res) <- rep("",nrow(res))
+  }
+  res
+}
+
+
+
+
+
+#' @title Confidence interval for Poisson counts.
+#'
+#' @description Computes a confidence interval for the Poisson counts.
+#'
+#' @details Computes a CI for the Poisson counts using the \code{exact}, gamma distribution (\code{daly}`), Byar's (\code{byar}), or normal approximation (\code{asymptotic}) methods.  This is largely a wrapper to \code{\link[epitools]{pois.exact}}, \code{\link[epitools]{pois.daly}}, \code{\link[epitools]{pois.byar}}, and \code{\link[epitools]{pois.approx}} functions in \pkg{epitools}.
+#'
+#' @param x A single number or vector that represents the number of observed successes.
+#' @param conf.level A number that indicates the level of confidence to use for constructing confidence intervals (default is \code{0.95}).
+#' @param type A string that identifies the type of method to use for the calculations.  See details.
+#' @param verbose A logical that indicates whether \code{x} should be included in the returned matrix (\code{=TRUE}) or not (\code{=FALSE}; DEFAULT).
+#'
+#' @return A #x2 matrix that contains the lower and upper confidence interval bounds as columns and, if \code{verbose=TRUE} \code{x}.
+#'
+#' @author Derek H. Ogle, \email{derek@@derekogle.com}
+#'
+#' @seealso See \code{\link[epitools]{pois.exact}}, \code{\link[epitools]{pois.daly}}, \code{\link[epitools]{pois.byar}}, and \code{\link[epitools]{pois.approx}} in \pkg{epitools} for more description and references.
+#'
+#' @keywords htest
+#'
+#' @examples
+#' ## Demonstrates using all types at once
+#' poiCI(12)
+#' 
+#' ## Selecting types
+#' poiCI(12,type="daly")
+#' poiCI(12,type="byar")
+#' poiCI(12,type="asymptotic")
+#' poiCI(12,type="asymptotic",verbose=TRUE)
+#' poiCI(12,type=c("exact","daly"))
+#' poiCI(12,type=c("exact","daly"),verbose=TRUE)
+#' 
+#' ## Demonstrates use with multiple inputs
+#' poiCI(c(7,10),type="exact")
+#' poiCI(c(7,10),type="exact",verbose=TRUE)
+#' 
+#' @export
+poiCI <- function(x,conf.level=0.95,type=c("exact","daly","byar","asymptotic"),
+                  verbose=FALSE) {
+  type <- match.arg(type,several.ok=TRUE)
+  if (!is.vector(x)) STOP("'x' must be a single or vector of whole numbers.")
+  if (!is.numeric(x)) STOP("'x' must be a whole number.")
+  if (!all(is.wholenumber(x))) STOP("'x' must be a whole number.")
+  if (any(x<0)) STOP("'x' must be non-negative.")
+  if (conf.level<=0 | conf.level>=1) STOP("'conf.level' must be between 0 and 1.")
+  ## Handle differently depending on number of xs given
+  if (length(x)>1) {
+    if (length(type)>1) {
+      type <- type[1]
+      WARN("Can't use multiple 'type's with multiple 'x's. Used only '",type,"'.")
+    }
+    switch(type,
+           exact = { res <- epitools::pois.exact(x,1,conf.level) },
+           daly  = { res <- epitools::pois.daly(x,1,conf.level) },
+           byar  = { res <- epitools::pois.byar(x,1,conf.level) },
+           asymptotic = { res <- epitools::pois.approx(x,1,conf.level) })
+  } else {
+    res <- rbind(epitools::pois.exact(x,1,conf.level),
+                 epitools::pois.daly(x,1,conf.level),
+                 epitools::pois.byar(x,1,conf.level),
+                 epitools::pois.approx(x,1,conf.level))
+    rownames(res) <- c("Exact","Daly","Byar","Asymptotic")
+    # reduce to selected types
+    res <- res[rownames(res) %in% capFirst(type),]
+  }
+  # relabel CI columns, convert to matrix, drop unwanted columns
+  names(res)[which(names(res) %in% c("lower","upper"))] <- iCILabel(conf.level)
+  res <- as.matrix(res[,-c(2,3,6)])
+  # return everything if verbose=TRUE, otherwise just CI
+  if (!verbose) {
+    res <- res[,2:3,drop=FALSE]
+    # remove rownname if only one type selected and not verbose
+    if (length(type)==1) rownames(res) <- rep("",nrow(res))
+  }
   res
 }
 
@@ -102,57 +199,20 @@ binCI <- function(x,n,conf.level=0.95,type=c("wilson","exact","asymptotic","all"
 #'
 #' @export
 hyperCI <- function(M,n,m,conf.level=0.95) {
-  if (!is.numeric(c(M,n,m))) STOP("'M', 'n', and 'm' must all be numeric.")
+  if (any(length(M)!=1,length(n)!=1,length(m)!=1)) STOP("'M','n', and 'm' must all be a single value.")
+  if (!is.numeric(c(M,n,m))) STOP("'M', 'n', and 'm' must all be whole numbers.")
+  if (!all(is.wholenumber(M))) STOP("'M' must be a whole number.")
+  if (!all(is.wholenumber(n))) STOP("'n' must be a whole number.")
+  if (!all(is.wholenumber(m))) STOP("'m' must be a whole number.")
   if (any(c(M,n,m)<1)) STOP("'M', 'n', and 'm' must all be non-negative.")
   if (m>n) STOP("'m' must be less than 'n'.")
   if (m>M) STOP("'m' must be less than 'M'.")
+  if (conf.level<=0 | conf.level>=1) STOP("'conf.level' must be between 0 and 1.")
   N.low <- (n+(M-m))
   while (stats::qhyper((1-conf.level)/2,n,N.low-n,M) > m) { N.low <- N.low + 1 }
   N.hi <- (n*M)/m
   while (stats::qhyper(1-((1-conf.level)/2),n,N.hi-n,M) >= m) { N.hi <- N.hi + 1 }
   res <- round(cbind(N.low,N.hi),0)
-  colnames(res) <- iCILabel(conf.level)
-  res
-}
-
-
-
-
-#' @title Confidence interval for Poisson rate parameter.
-#'
-#' @description Computes a confidence interval for the Poisson mean rate parameter.
-#'
-#' @details Computes a CI for the Poisson mean using the method described in Ulm (1990), though this method was earlier described by Liddell (1984) and possibly Garwood (1936) as noted in van der Gulden and Verbeck (1992).  Thank you to Jerry Lewis for clarifications to the historical citations of this method.
-#'
-#' @param x A number representing the number of observed successes.
-#' @param conf.level A number that indicates the level of confidence to use for constructing confidence intervals (default is \code{0.95}).
-#'
-#' @return A 1x2 matrix that contains the lower and upper confidence interval bounds.
-#'
-#' @author Derek H. Ogle, \email{derek@@derekogle.com}
-#'
-#' @seealso See \code{\link[epitools]{pois.exact}} in \pkg{epitools} for similar functionality.
-#'
-#' @references Garwood, F.  1936.  Fiducial limits for the Poisson distribution.  Biometrika.  28(3/4):437-442.
-#'
-#' Liddell, F.D.  1984.  Simple exact analysis of the standardised mortality ratio.  Journal of Epidemiology and Community Health. 38(1):85-88.
-#'
-#' Ulm, K.  1990.  A simple method to calculate the confidence interval of a standardized mortality ratio.  American Journal of Epidemiology 131(2):373-375.
-#'
-#' vand der Gulden, J.W.J. and A.L.M. Verbeck.  1992.  Re: \dQuote{A simple method to calculate the confidence interval of a standardized mortality ratio (SMR)}.  American Journal of Epidemiology 136(9):1170-1171.
-#'
-#' @keywords htest
-#'
-#' @examples
-#' poiCI(12)
-#'
-#' @export
-poiCI <- function(x,conf.level=0.95) {
-  if (!is.numeric(x)) STOP("'x' must be numeric.")
-  if (x<1) STOP("'x' must be non-negative.")
-  LCI <- stats::qchisq((1-conf.level)/2,2*x)/2
-  UCI <- stats::qchisq(1-(1-conf.level)/2,2*(x+1))/2
-  res <- cbind(LCI,UCI)
   colnames(res) <- iCILabel(conf.level)
   res
 }